address PR feedback
- move replicate to space
- add code comments based on RFC.
