+++ capabilities/blob/blob.ipldsch
@@ -30,9 +30,3 @@ type AcceptOk struct {
   site Link
   PDP optional Link
 }
-
-type ReplicateCaveats struct {
-  blob Blob
-  replicas Int
-  location Link
-}
