+++ capabilities/replica/allocate.go
@@ -8,7 +8,7 @@ import (
 	"github.com/storacha/go-ucanto/ucan"
 	"github.com/storacha/go-ucanto/validator"
 
-	"github.com/storacha/go-libstoracha/capabilities/blob"
+	"github.com/storacha/go-libstoracha/capabilities/space/blob"
 	"github.com/storacha/go-libstoracha/capabilities/types"
 )
 
@@ -17,13 +17,19 @@ const AllocateAbility = "replica/allocate"
 var _ ipld.Builder = (*AllocateCaveats)(nil)
 
 type AllocateCaveats struct {
-	Space    did.DID
-	Blob     blob.Blob
+	// Space contains the did to allocate Blob in.
+	Space did.DID
+	// Blob is the blob to be allocated.
+	Blob blob.Blob
+	// Location contains a location commitment indicating where the Blob must be
+	// fetched from.
 	Location ucan.Link
-	Cause    ucan.Link
+	// Cause contains the `space/blob/replicate` invocation that caused this allocation.
+	Cause ucan.Link
 }
 
 type AllocateOk struct {
+	// Size is the number of bytes allocated for a Blob.
 	Size uint64
 }
 
@@ -36,6 +42,13 @@ func (ac AllocateCaveats) ToIPLD() (datamodel.Node, error) {
 }
 
 var AllocateCaveatsReader = schema.Struct[AllocateCaveats](AllocateCaveatsType(), nil, types.Converters...)
+
+// Allocate is a capability that allows an agent to allocate a Blob for replication
+// into a space identified by did:key in the `with` field.
+//
+// The Allocate task receipt includes an async task that will be performed by
+// a storage node - `blob/replica/transfer`. The `blob/replica/transfer` task is
+// completed when the storage node has transferred the blob from its location to the storage node.
 var Allocate = validator.NewCapability(
 	AllocateAbility,
 	schema.DIDString(),
