+++ capabilities/replica/transfer.go
@@ -9,7 +9,7 @@ import (
 	"github.com/storacha/go-ucanto/ucan"
 	"github.com/storacha/go-ucanto/validator"
 
-	"github.com/storacha/go-libstoracha/capabilities/blob"
+	"github.com/storacha/go-libstoracha/capabilities/space/blob"
 	"github.com/storacha/go-libstoracha/capabilities/types"
 )
 
@@ -18,15 +18,24 @@ const TransferAbility = "replica/transfer"
 var _ ipld.Builder = (*TransferCaveats)(nil)
 
 type TransferCaveats struct {
-	Space    did.DID
-	Blob     blob.Blob
+	// Space contains the did to transfer Blob to.
+	Space did.DID
+	// Blob is the blob to be transferred.
+	Blob blob.Blob
+	// Location contains a location commitment indicating where the Blob must be
+	// transferred from.
 	Location ucan.Link
-	Cause    ucan.Link
+	// Cause contains the `blob/replica/allocate` invocation that initiated this transfer.
+	Cause ucan.Link
 }
 
 type TransferOk struct {
+	// Site contains the location commitment indicate where the Blob has been
+	// transferred to.
 	Site ucan.Link
-	PDP  *ucan.Link
+	// PDP optionally contains the PDP invocation will complete when aggregation
+	// is complete and the piece is accepted.
+	PDP *ucan.Link
 }
 
 func (t TransferOk) ToIPLD() (datamodel.Node, error) {
@@ -38,6 +47,9 @@ func (tc TransferCaveats) ToIPLD() (datamodel.Node, error) {
 }
 
 var TransferCaveatsReader = schema.Struct[TransferCaveats](TransferCaveatsType(), nil, types.Converters...)
+
+// Transfer is a capability that allows an agent to transfer a Blob for replication
+// into a space identified by did:key in the `with` field.
 var Transfer = validator.NewCapability(
 	TransferAbility,
 	schema.DIDString(),
